moved and renamed OpCat.svelte
