+++ src/components/controls/RasterOpCat.svelte
